Universal session keep alive in employee front-end
