+++ frontend/src/employee-frontend/components/EmployeeRoute.tsx
@@ -5,11 +5,15 @@
 import React, { useContext, useEffect } from 'react'
 import { Navigate } from 'react-router'
 
+import SessionExpiredModal from 'lib-components/molecules/modals/SessionExpiredModal'
+import { useKeepSessionAlive } from 'lib-components/useKeepSessionAlive'
 import { Translations } from 'lib-customizations/employee'
 
 import { useTranslation } from '../state/i18n'
 import { TitleContext, TitleState } from '../state/title'
 import { UserContext } from '../state/user'
+
+import { sessionKeepalive } from './common/sessionKeepalive'
 
 interface Props {
   title?: keyof Translations['titles']
@@ -40,6 +44,19 @@ const RequireAuth = React.memo(function EnsureAuthenticated({
 }: {
   element: React.ReactNode
 }) {
+  const { showSessionExpiredModal, setShowSessionExpiredModal } =
+    useKeepSessionAlive(sessionKeepalive)
   const { loggedIn } = useContext(UserContext)
-  return loggedIn ? <>{element}</> : <Navigate replace to="/" />
+  return loggedIn ? (
+    <>
+      {element}
+      {showSessionExpiredModal && (
+        <SessionExpiredModal
+          onClose={() => setShowSessionExpiredModal(false)}
+        />
+      )}
+    </>
+  ) : (
+    <Navigate replace to="/" />
+  )
 })
